requirements.txt
fair_esm==2.0.0
imblearn==0.0
joblib==1.4.2
lightgbm==4.5.0
matplotlib==3.7.2
matplotlib==3.7.5
mlxtend==0.22.0
numpy==1.24.4
pandas==1.5.0
pyfasta==0.5.2
rich==13.9.2
scikit_learn==1.3.2
seaborn==0.13.2
torch==1.8.2+cu111
torchvision==0.9.2+cu111
tqdm==4.66.5
umap==0.1.1
